requirements.txt
# FastAPI 및 서버 관련
fastapi==0.104.1
uvicorn==0.23.2
websockets==11.0.3
httpx==0.25.1
pydantic==2.4.2

# 테스트 관련
pytest==7.4.3
pytest-asyncio==0.21.1

# 유틸리티
python-dotenv==1.0.0 